ResNet/requirements.txt
torch
torchvision
numpy
Pillow
six
pyyaml
tqdm